requirements.txt
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
coinbase==2.1.0
constantly==15.1.0
cryptography==3.3.1
dateparser==1.0.0
et-xmlfile==1.0.1
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
jdcal==1.4.1
numpy==1.19.5
openpyxl==3.0.6
pandas==1.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.9
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
python-binance==0.7.9
python-dateutil==2.8.1
pytz==2020.5
regex==2020.11.13
requests==2.25.1
service-identity==18.1.0
six==1.15.0
Twisted==20.3.0
txaio==20.12.1
tzlocal==2.1
ujson==4.0.1
urllib3==1.26.2
XlsxWriter==1.3.7
zope.interface==5.2.0
python-dotenv 
